// src/app/models/gallery-item.model.ts
export interface GalleryItem {
  id: number;
  title: string;
  titleMr: string;
  titleEn: string;
  titleHi: string;
  description: string;
  descriptionMr: string;
  descriptionEn: string;
  descriptionHi: string;
  imageUrl: string;
  category: string;
  uploadDate: Date;
  isActive: boolean;
}